// src/data/currency.d.ts
// Lowercase currency codes
export type CurrencyCode =
  | "aed"
  | "afn"
  | "all"
  | "amd"
  | "ang"
  | "aoa"
  | "ars"
  | "aud"
  | "awg"
  | "azn"
  | "bam"
  | "bbd"
  | "bdt"
  | "bgn"
  | "bhd"
  | "bif"
  | "bmd"
  | "bnd"
  | "bob"
  | "brl"
  | "bsd"
  | "btn"
  | "bwp"
  | "byn"
  | "bzd"
  | "cad"
  | "cdf"
  | "chf"
  | "clp"
  | "cnh"
  | "cny"
  | "cop"
  | "crc"
  | "cuc"
  | "cup"
  | "cve"
  | "czk"
  | "djf"
  | "dkk"
  | "dop"
  | "dzd"
  | "egp"
  | "ern"
  | "etb"
  | "eur"
  | "fjd"
  | "fkp"
  | "gbp"
  | "gel"
  | "ggp"
  | "ghs"
  | "gip"
  | "gmd"
  | "gnf"
  | "gtq"
  | "gyd"
  | "hkd"
  | "hnl"
  | "htg"
  | "huf"
  | "idr"
  | "ils"
  | "imp"
  | "inr"
  | "iqd"
  | "irr"
  | "isk"
  | "jep"
  | "jmd"
  | "jod"
  | "jpy"
  | "kes"
  | "kgs"
  | "khr"
  | "kmf"
  | "kpw"
  | "krw"
  | "kwd"
  | "kyd"
  | "kzt"
  | "lak"
  | "lbp"
  | "lkr"
  | "lrd"
  | "lsl"
  | "lyd"
  | "mad"
  | "mdl"
  | "mga"
  | "mkd"
  | "mmk"
  | "mnt"
  | "mop"
  | "mru"
  | "mur"
  | "mvr"
  | "mwk"
  | "mxn"
  | "myr"
  | "mzn"
  | "nad"
  | "ngn"
  | "nio"
  | "nok"
  | "npr"
  | "nzd"
  | "omr"
  | "pab"
  | "pen"
  | "pgk"
  | "php"
  | "pkr"
  | "pln"
  | "pyg"
  | "qar"
  | "ron"
  | "rsd"
  | "rub"
  | "rwf"
  | "sar"
  | "sbd"
  | "scr"
  | "sdg"
  | "sek"
  | "sgd"
  | "shp"
  | "sos"
  | "srd"
  | "stn"
  | "svc"
  | "syp"
  | "szl"
  | "thb"
  | "tjs"
  | "tmt"
  | "tnd"
  | "top"
  | "try"
  | "ttd"
  | "tvd"
  | "twd"
  | "tzs"
  | "uah"
  | "ugx"
  | "usd"
  | "uyu"
  | "uzs"
  | "vnd"
  | "vuv"
  | "wst"
  | "xaf"
  | "xcd"
  | "xdr"
  | "xof"
  | "xpf"
  | "yer"
  | "zar"
  | "zmw";

// Denomination (number of decimals)
export type CurrencyDenomination = 0 | 1 | 2 | 3 | 4;

// ISO 639-1
type LocaleCode =
  `${Lowercase<string & `${Uppercase<string>[0]}${Uppercase<string>[1]}`>}-${Uppercase<string & `${Uppercase<string>[0]}${Uppercase<string>[1]}`>}`;

export type Currency = {
  currencyCode: CurrencyCode;
  englishCurrencyName: string;
  localCurrencyName: string;
  currencySymbol: string;
  countryRegionFlagEmoji: string;
  languageCode: LocaleCode;
  smallestCommonDenomination: number;
  acceptedCountries?: string[];
};